Move wireframe endpoint records to start package
