+++ backend/ads-service/src/main/java/com/marketinghub/geralanding/wireframe/service/start/GeraLandingWireframeExecutionSummaryResponse.java
@@ -1,4 +1,4 @@
-package com.marketinghub.geralanding.wireframe.service;
+package com.marketinghub.geralanding.wireframe.service.start;
 
 import java.math.BigDecimal;
 import java.time.Instant;
