+++ backend/ads-service/src/main/java/com/marketinghub/geralanding/wireframe/service/start/GeraLandingWireframeStartResponse.java
@@ -1,4 +1,4 @@
-package com.marketinghub.geralanding.wireframe.service;
+package com.marketinghub.geralanding.wireframe.service.start;
 
 /** Representa a resposta de início da etapa de wireframe do GeraLanding. */
 public record GeraLandingWireframeStartResponse(String idJob, String status) {}
