+++ backend/ads-service/src/main/java/com/marketinghub/geralanding/wireframe/service/start/RecordWireframeExperiment.java
@@ -1,4 +1,4 @@
-package com.marketinghub.geralanding.wireframe.service;
+package com.marketinghub.geralanding.wireframe.service.start;
 
 /** Representa os dados do experimento expostos na fila interna da etapa wireframe. */
 public record RecordWireframeExperiment(
