+++ backend/ads-service/src/main/java/com/marketinghub/geralanding/wireframe/service/start/RecordWireframePending.java
@@ -1,4 +1,4 @@
-package com.marketinghub.geralanding.wireframe.service;
+package com.marketinghub.geralanding.wireframe.service.start;
 
 /** Representa o item mínimo de pendência da etapa wireframe consumido pelo Worker AI. */
 public record RecordWireframePending(
