+++ backend/ads-service/src/test/java/com/marketinghub/geralanding/wireframe/service/BackendWireframeServiceTest.java
@@ -14,6 +14,8 @@ import com.marketinghub.experiment.Experiment;
 import com.marketinghub.experiment.repository.ExperimentRepository;
 import com.marketinghub.geralanding.GeraLandingStageExecution;
 import com.marketinghub.geralanding.GeraLandingStageExecutionRepository;
+import com.marketinghub.geralanding.wireframe.service.start.GeraLandingWireframeStartResponse;
+import com.marketinghub.geralanding.wireframe.service.start.RecordWireframePending;
 import java.nio.charset.StandardCharsets;
 import java.time.Instant;
 import java.util.List;
