+++ backend/ads-service/src/test/java/com/marketinghub/geralanding/wireframe/web/BackendWireframeControllerTest.java
@@ -10,10 +10,10 @@ import static org.mockito.Mockito.when;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.marketinghub.geralanding.wireframe.service.BackendWireframeService;
-import com.marketinghub.geralanding.wireframe.service.GeraLandingWireframeStartResponse;
-import com.marketinghub.geralanding.wireframe.service.RecordWireframeExperiment;
-import com.marketinghub.geralanding.wireframe.service.RecordWireframeHypothesis;
-import com.marketinghub.geralanding.wireframe.service.RecordWireframePending;
+import com.marketinghub.geralanding.wireframe.service.start.GeraLandingWireframeStartResponse;
+import com.marketinghub.geralanding.wireframe.service.start.RecordWireframeExperiment;
+import com.marketinghub.geralanding.wireframe.service.start.RecordWireframeHypothesis;
+import com.marketinghub.geralanding.wireframe.service.start.RecordWireframePending;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
